pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "payment-gateway"
version = "0.1.0"
description = "A modern payment gateway supporting both traditional banking and cryptocurrency payments"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "pydantic==2.6.1",
    "sqlalchemy==2.0.27",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "blockcypher==1.0.93",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "pytest==8.0.0",
    "httpx==0.26.0",
    "pydantic-settings==2.1.0",
    "loguru>=0.7.3",
    "pytest>=6.2.5",
    "pytest-mock>=3.6.1" 
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
namespaces = false

[tool.setuptools]
package-dir = {"" = "."}

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"] 